backend/MALEYANGA/requirements.txt
requests
python-dotenv
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
